+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.1 (2026-02-10)
+
+- **[Feature]** Add the following methods to `Extent`: `end`, `strip_start_checked`, `strip_end_checked`, `pop_start`, `pop_end`.
+- **[Feature]** Add the following methods to `XsliceRef`: `split_first_checked`, `sub_slice_to_checked`, `sub_slice_from_checked`, `const_copy`, `reborrow.
+- **[Feature]** Add the following methods to `XbstrRef`: `split_first_checked`, `sub_slice_to_checked`, `sub_slice_from_checked`, `const_copy`, `reborrow.
+- **[Feature]** Add the following methods to `XstrRef`: `sub_slice_checked`.
+- **[Feature]** Add support for mutable slices, through `XsliceMut`.
+- **[Feature]** Add support for reborrowing of `XsliceRef` and `XsliceMut`.
+
 # 0.1.0 (2026-02-09)
 
 - **[Feature]** First version